pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawai"
version = "2.0.0"
description = "AI-powered automated penetration testing system integrating LLM agents with security tools for vulnerability discovery, exploitation, and reporting"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "ClawAI Team"},
]
keywords = ["penetration-testing", "security", "ai", "llm", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Information Technology",
    "Topic :: Security",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web framework
    "fastapi>=0.104.1",
    "python-multipart>=0.0.6",
    "uvicorn[standard]>=0.24.0",
    # CLI and TUI
    "click>=8.0.0",
    "rich>=13.0.0",
    "textual>=0.40.0",
    # Auth and security
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    # Data validation and serialization
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Database and ORM
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.1",
    "redis>=5.0.1",
    # HTTP and networking
    "httpx>=0.25.1",
    "requests>=2.31.0",
    # Async
    "aiofiles>=23.2.1",
    # Config and templating
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    # AI
    "openai>=1.0.0",
    # Data validation
    "jsonschema>=4.19.2",
    # Search
    "duckduckgo-search>=4.0.0",
    # Monitoring
    "prometheus-client>=0.19.0",
    # System utilities
    "psutil>=5.9.6",
    "cryptography>=41.0.0",
    # WebSocket
    "websockets>=12.0",
    # Container management
    "docker>=6.1.3",
    # Typing
    "typing-extensions>=4.9.0",
]

[project.optional-dependencies]
ml = [
    "sentence-transformers>=2.2.0",
    "qdrant-client>=1.6.0",
    "numpy>=1.24.4",
    "pandas>=2.1.4",
]
neo4j = [
    "neo4j>=5.20.0",
    "pgvector>=0.2.0",
]
postgres = [
    "psycopg2-binary>=2.9.9",
]
flask = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "pytest-mock>=3.10",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "bandit>=1.7",
    "pip-audit>=2.6",
    "pre-commit>=3.0",
]

[project.scripts]
clawai = "src.cli.main:cli"

[tool.setuptools.packages.find]
include = ["src*"]

# ==================== Pytest ====================
[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "perf: marks performance tests",
    "integration: marks integration tests that require external services",
    "unit: marks unit tests",
]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

# ==================== Coverage ====================
[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "scripts/*",
    "tools/*",
    "references/*",
]

[tool.coverage.report]
fail_under = 30
show_missing = true

# ==================== Ruff ====================
[tool.ruff]
line-length = 120
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "S", "B", "A", "PIE", "Q", "T20"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"scripts/*" = ["T20"]

# ==================== Mypy ====================
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
